alert-consumer/requirements.txt
confluent-kafka[avro]==2.3.0
redis==5.0.1
psycopg2-binary==2.9.9
requests==2.31.0
structlog==24.1.0
python-dotenv==1.0.1
